requirements.txt
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
certifi==2024.8.30
distro==1.9.0
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.2
flake8==7.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.0
mccabe==0.7.0
openai==1.57.0
psycopg2-binary==2.9.9
pycodestyle==2.12.1
pydantic==2.10.3
pydantic_core==2.27.1
pyflakes==3.2.0
python-dateutil==2.9.0.post0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.1
tqdm==4.67.1
typing_extensions==4.12.2
